+++ src/Rendering/TextureOverlayRenderer.hpp
@@ -37,12 +37,12 @@ class TextureOverlayRenderer : public IVistaOpenGLDraw {
   /**
    * Set the opacity of the overlay
    */
-  void SetOpacity(double val);
+  void SetOpacity(float val);
 
   /**
    * Set the time value passed to shader to discard
    */
-  void SetTime(double val);
+  void SetTime(float val);
 
   /**
    * Set if timing information should be used by the shader
